Add work item templates operations and enhance types module documentation
