+++ src/mcp_azure_devops/features/work_items/tools/types.py
@@ -1,16 +1,23 @@
 """
-Work item types and metadata operations for Azure DevOps.
+Work item types and fields operations for Azure DevOps.
 
-This module provides MCP tools for retrieving work item types, templates, and fields.
+This module provides MCP tools for retrieving work item types and fields.
 """
-from typing import Optional
-
 from azure.devops.v7_1.work_item_tracking import WorkItemTrackingClient
 
 from mcp_azure_devops.features.work_items.common import (
     AzureDevOpsClientError,
     get_work_item_client,
 )
+
+
+def _format_table(headers, rows):
+    """Format data as a markdown table."""
+    result = []
+    result.append("| " + " | ".join(headers) + " |")
+    result.append("| " + " | ".join(["----"] * len(headers)) + " |")
+    result.extend(rows)
+    return "\n".join(result)
 
 
 def _format_work_item_type(wit):
@@ -34,7 +41,7 @@ def _format_work_item_type(wit):
     if states:
         result.append("\n## States")
         for state in states:
-            state_info = f"- **{state.name}** (Category: {state.state_category}, Color: {state.color})"
+            state_info = f"- **{state.name}** (Category: {state.category}, Color: {state.color})"
             order = getattr(state, "order", None)
             if order is not None:
                 state_info += f", Order: {order}"
@@ -70,66 +77,27 @@ def _format_work_item_field(field):
     return "\n".join(result)
 
 
-def _format_work_item_template(template):
-    """Format work item template data for display."""
-    result = [f"# Template: {template.name}"]
-    
-    for attr in ["description", "work_item_type_name", "id"]:
-        value = getattr(template, attr, None)
-        if value:
-            result.append(f"**{attr.replace('_', ' ').capitalize()}:** {value}")
-    
-    fields = getattr(template, "fields", None)
-    if fields:
-        result.append("\n## Default Field Values")
-        for field, value in fields.items():
-            result.append(f"- **{field}**: {value}")
-    
-    return "\n".join(result)
-
-
 def _get_work_item_types_impl(project: str, wit_client: WorkItemTrackingClient) -> str:
-    """
-    Implementation of work item types retrieval.
-    
-    Args:
-        project: Project name or ID
-        wit_client: Work item tracking client
-    
-    Returns:
-        Formatted string containing work item types information
-    """
+    """Implementation of work item types retrieval."""
     work_item_types = wit_client.get_work_item_types(project)
     
     if not work_item_types:
         return f"No work item types found in project {project}."
     
-    # Format simple list overview first
-    result = [f"# Work Item Types in Project: {project}\n"]
-    result.append("| Name | Reference Name | Description |")
-    result.append("| ---- | -------------- | ----------- |")
+    headers = ["Name", "Reference Name", "Description"]
     
-    for wit in work_item_types:
-        description = getattr(wit, "description", "N/A")
-        ref_name = getattr(wit, "reference_name", "N/A")
-        result.append(f"| {wit.name} | {ref_name} | {description} |")
+    # Use list comprehension for more concise table building
+    rows = [
+        f"| {wit.name} | {getattr(wit, 'reference_name', 'N/A')} | {getattr(wit, 'description', 'N/A')} |"
+        for wit in work_item_types
+    ]
     
-    return "\n".join(result)
+    return f"# Work Item Types in Project: {project}\n\n" + _format_table(headers, rows)
 
 
 def _get_work_item_type_impl(project: str, type_name: str, 
                              wit_client: WorkItemTrackingClient) -> str:
-    """
-    Implementation of work item type detail retrieval.
-    
-    Args:
-        project: Project name or ID
-        type_name: Work item type name
-        wit_client: Work item tracking client
-    
-    Returns:
-        Formatted string containing work item type details
-    """
+    """Implementation of work item type detail retrieval."""
     work_item_type = wit_client.get_work_item_type(project, type_name)
     
     if not work_item_type:
@@ -140,53 +108,35 @@ def _get_work_item_type_impl(project: str, type_name: str,
 
 def _get_work_item_type_fields_impl(project: str, type_name: str, 
                                     wit_client: WorkItemTrackingClient) -> str:
-    """
-    Implementation of work item type fields retrieval.
-    
-    Args:
-        project: Project name or ID
-        type_name: Work item type name
-        wit_client: Work item tracking client
-    
-    Returns:
-        Formatted string containing work item type fields
-    """
+    """Implementation of work item type fields retrieval."""
     try:
         fields = wit_client.get_work_item_type_fields_with_references(project, type_name, expand="all")
         
         if not fields:
             return f"No fields found for work item type '{type_name}' in project {project}."
         
-        # Format simple list overview first
-        result = [f"# Fields for Work Item Type: {type_name}\n"]
-        result.append("| Name | Reference Name | Type | Required | Read Only |")
-        result.append("| ---- | -------------- | ---- | -------- | --------- |")
+        headers = ["Name", "Reference Name", "Type", "Required", "Read Only", "Allowed Values", "Default Value", "Help Text"]
         
-        for field in fields:
-            required = "Yes" if getattr(field, "always_required", False) else "No"
-            read_only = "Yes" if getattr(field, "read_only", False) else "No"
-            field_type = getattr(field, "type", "N/A")
-            result.append(f"| {field.name} | {field.reference_name} | {field_type} | {required} | {read_only} |")
+        # Use list comprehension for table rows
+        rows = [
+            f"| {field.name} | {field.reference_name} | " +
+            f"{getattr(field, 'type', 'N/A')} | " +
+            f"{'Yes' if getattr(field, 'always_required', False) else 'No'} | " +
+            f"{'Yes' if getattr(field, 'read_only', False) else 'No'} |" +
+            f"{', '.join(getattr(field, 'allowed_values', []))} | " +
+            f"{getattr(field, 'default_value', 'N/A')} |" +
+            f"{getattr(field, 'help_text', 'N/A')} |"
+            for field in fields
+        ]
         
-        return "\n".join(result)
+        return f"# Fields for Work Item Type: {type_name}\n\n" + _format_table(headers, rows)
     except Exception as e:
         return f"Error retrieving fields for work item type '{type_name}' in project '{project}': {str(e)}"
 
 
 def _get_work_item_type_field_impl(project: str, type_name: str, field_name: str,
                                    wit_client: WorkItemTrackingClient) -> str:
-    """
-    Implementation of work item type field detail retrieval.
-    
-    Args:
-        project: Project name or ID
-        type_name: Work item type name
-        field_name: Field reference name or display name
-        wit_client: Work item tracking client
-    
-    Returns:
-        Formatted string containing work item type field details
-    """
+    """Implementation of work item type field detail retrieval."""
     try:
         field = wit_client.get_work_item_type_field_with_references(
             project, type_name, field_name, expand="all")
@@ -199,87 +149,9 @@ def _get_work_item_type_field_impl(project: str, type_name: str, field_name: str
         return f"Error retrieving field '{field_name}' for work item type '{type_name}' in project '{project}': {str(e)}"
 
 
-def _create_team_context(team_context_dict):
-    """Create a TeamContext object from a dictionary."""
-    from azure.devops.v7_1.work_item_tracking.models import TeamContext
-    return TeamContext(
-        project=team_context_dict.get('project'),
-        project_id=team_context_dict.get('project_id'),
-        team=team_context_dict.get('team'),
-        team_id=team_context_dict.get('team_id')
-    )
-
-
-def _get_work_item_templates_impl(team_context: dict, work_item_type: Optional[str] ,
-                                 wit_client: WorkItemTrackingClient) -> str:
-    """
-    Implementation of work item templates retrieval.
-    
-    Args:
-        team_context: Team context dictionary with project and team information
-        work_item_type: Optional work item type to filter templates
-        wit_client: Work item tracking client
-    
-    Returns:
-        Formatted string containing work item templates
-    """
-    try:
-        team_ctx = _create_team_context(team_context)
-        templates = wit_client.get_templates(team_ctx, work_item_type)
-        
-        if not templates:
-            filter_msg = f" for work item type '{work_item_type}'" if work_item_type else ""
-            return f"No templates found{filter_msg} in team {team_context.get('team') or team_context.get('team_id')}."
-        
-        # Format simple list overview first
-        project_display = team_context.get('project') or team_context.get('project_id')
-        team_display = team_context.get('team') or team_context.get('team_id')
-        result = [f"# Work Item Templates for Team: {team_display} (Project: {project_display})\n"]
-        
-        if work_item_type:
-            result[0] += f" (Filtered by type: {work_item_type})"
-        
-        result.append("| Name | Work Item Type | Description |")
-        result.append("| ---- | -------------- | ----------- |")
-        
-        for template in templates:
-            description = getattr(template, "description", "N/A")
-            wit_name = getattr(template, "work_item_type_name", "N/A")
-            result.append(f"| {template.name} | {wit_name} | {description} |")
-        
-        return "\n".join(result)
-    except Exception as e:
-        return f"Error retrieving templates: {str(e)}"
-
-
-def _get_work_item_template_impl(team_context: dict, template_id: str,
-                                wit_client: WorkItemTrackingClient) -> str:
-    """
-    Implementation of work item template detail retrieval.
-    
-    Args:
-        team_context: Team context dictionary with project and team information
-        template_id: Template ID
-        wit_client: Work item tracking client
-    
-    Returns:
-        Formatted string containing work item template details
-    """
-    try:
-        team_ctx = _create_team_context(team_context)
-        template = wit_client.get_template(team_ctx, template_id)
-        
-        if not template:
-            return f"Template with ID '{template_id}' not found."
-        
-        return _format_work_item_template(template)
-    except Exception as e:
-        return f"Error retrieving template '{template_id}': {str(e)}"
-
-
 def register_tools(mcp) -> None:
     """
-    Register work item types and metadata tools with the MCP server.
+    Register work item types and fields tools with the MCP server.
     
     Args:
         mcp: The FastMCP server instance
@@ -378,60 +250,5 @@ def register_tools(mcp) -> None:
         try:
             wit_client = get_work_item_client()
             return _get_work_item_type_field_impl(project, type_name, field_name, wit_client)
-        except AzureDevOpsClientError as e:
-            return f"Error: {str(e)}"
-    
-    @mcp.tool()
-    def get_work_item_templates(team_context: dict, work_item_type: Optional[str]) -> str:
-        """
-        Gets a list of all work item templates for a team.
-        
-        Use this tool when you need to:
-        - Find available templates for creating work items
-        - Get template IDs for use in other operations
-        - Filter templates by work item type
-        
-        Args:
-            team_context: Dictionary containing team information with keys:
-                project: Project name (Optional if project_id is provided)
-                project_id: Project ID (Optional if project is provided)
-                team: Team name (Optional if team_id is provided)
-                team_id: Team ID (Optional if team is provided)
-            work_item_type: Optional work item type name to filter templates
-            
-        Returns:
-            A formatted table of all templates with names, work item types,
-            and descriptions
-        """
-        try:
-            wit_client = get_work_item_client()
-            return _get_work_item_templates_impl(team_context, work_item_type, wit_client)
-        except AzureDevOpsClientError as e:
-            return f"Error: {str(e)}"
-    
-    @mcp.tool()
-    def get_work_item_template(team_context: dict, template_id: str) -> str:
-        """
-        Gets detailed information about a specific work item template.
-        
-        Use this tool when you need to:
-        - View default field values in a template
-        - Understand what a template pre-populates in a work item
-        - Get complete details about a template
-        
-        Args:
-            team_context: Dictionary containing team information with keys:
-                project: Project name (Optional if project_id is provided)
-                project_id: Project ID (Optional if project is provided)
-                team: Team name (Optional if team_id is provided)
-                team_id: Team ID (Optional if team is provided)
-            template_id: The ID of the template
-            
-        Returns:
-            Detailed information about the template including default field values
-        """
-        try:
-            wit_client = get_work_item_client()
-            return _get_work_item_template_impl(team_context, template_id, wit_client)
         except AzureDevOpsClientError as e:
             return f"Error: {str(e)}"
